Initial commit43

+++ app/src/main/java/com/dippola/relaxtour/setting/StorageManageAdapter.java
@@ -1,7 +1,10 @@
 package com.dippola.relaxtour.setting;
 
+import static android.content.Context.MODE_PRIVATE;
+
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -22,7 +25,10 @@ import com.dippola.relaxtour.controller.RainController;
 import com.dippola.relaxtour.notification.NotificationService;
 import com.dippola.relaxtour.pages.ChakraPage;
 import com.dippola.relaxtour.pages.HzPage;
+import com.dippola.relaxtour.pages.MantraPage;
+import com.dippola.relaxtour.pages.NaturePage;
 import com.dippola.relaxtour.pages.RainPage;
+import com.dippola.relaxtour.pages.WaterPage;
 import com.dippola.relaxtour.pages.WindPage;
 import com.dippola.relaxtour.pages.adapter.FavListAdapter;
 
@@ -54,20 +60,26 @@ public class StorageManageAdapter extends  RecyclerView.Adapter<StorageManageAda
         holder.deleteBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                StorageManageDialog.progressBar.setVisibility(View.VISIBLE);
-                String path = context.getApplicationInfo().dataDir + "/cache/" + arrayList.get(i).getName() + ".mp3";
-                File file = new File(path);
-                if (file.exists()) {
-                    file.delete();
-                    StorageManageDialogItem storageManageDialogItem = arrayList.get(i);
-                    int index = arrayList.indexOf(arrayList.get(i));
-                    arrayList.remove(index);
-                    notifyItemRemoved(index);
-                    notifyDataSetChanged();
-                    Log.d(">>>MainActivity", "deleted");
-                    resetPage(storageManageDialogItem);
+                SharedPreferences sharedPreferences = context.getSharedPreferences("storage_manage_checkbox", MODE_PRIVATE);
+                boolean isChecked = sharedPreferences.getBoolean("isChecked", false);
+                if (isChecked) {
+                    StorageManageDialog.progressBar.setVisibility(View.VISIBLE);
+                    String path = context.getApplicationInfo().dataDir + "/cache/" + "audio" + arrayList.get(i).getPage() + "to" + arrayList.get(i).getPosition() + ".mp3";
+                    File file = new File(path);
+                    if (file.exists()) {
+                        file.delete();
+                        StorageManageDialogItem storageManageDialogItem = arrayList.get(i);
+                        int index = arrayList.indexOf(arrayList.get(i));
+                        arrayList.remove(index);
+                        notifyItemRemoved(index);
+                        notifyDataSetChanged();
+                        Log.d(">>>MainActivity", "deleted");
+                        resetPage(storageManageDialogItem, context);
+                    } else {
+                        Log.d(">>>MainActivity", "no have");
+                    }
                 } else {
-                    Log.d(">>>MainActivity", "no have");
+                    StorageManageDialogAskSure.storageManageDialogAskSure(context, arrayList, StorageManageAdapter.this, i);
                 }
             }
         });
@@ -88,16 +100,26 @@ public class StorageManageAdapter extends  RecyclerView.Adapter<StorageManageAda
         }
     }
 
-    private void resetPage(StorageManageDialogItem storageManageDialogItem) {
-        if (storageManageDialogItem.getPage() == 3) {
+    public static void resetPage(StorageManageDialogItem storageManageDialogItem, Context context) {
+        if (storageManageDialogItem.getPage() == 1 && RainPage.arrayList.size() != 0) {
+            RainPage.adapter.notifyDataSetChanged();
+        } else if (storageManageDialogItem.getPage() == 2 && WaterPage.arrayList.size() != 0) {
+            WaterPage.adapter.notifyDataSetChanged();
+        } else if (storageManageDialogItem.getPage() == 3 && WindPage.arrayList.size() != 0) {
+            WindPage.adapter.notifyDataSetChanged();
+        } else if (storageManageDialogItem.getPage() == 4 && NaturePage.arrayList.size() != 0) {
+            NaturePage.adapter.notifyDataSetChanged();
+        } else if (storageManageDialogItem.getPage() == 5 && ChakraPage.arrayList.size() != 0) {
             ChakraPage.adapter.notifyDataSetChanged();
-        } else if (storageManageDialogItem.getPage() == 4) {
+        } else if (storageManageDialogItem.getPage() == 6 && MantraPage.arrayList.size() != 0) {
+            MantraPage.adapter.notifyDataSetChanged();
+        } else if (storageManageDialogItem.getPage() == 7 && HzPage.arrayList.size() != 0) {
             HzPage.adapter.notifyDataSetChanged();
         }
-        deleteInBottomPlayList(storageManageDialogItem);
+        deleteInBottomPlayList(storageManageDialogItem, context);
     }
 
-    private void deleteInBottomPlayList(StorageManageDialogItem storageManageDialogItem) {
+    public static void deleteInBottomPlayList(StorageManageDialogItem storageManageDialogItem, Context context) {
         if (MainActivity.bottomSheetPlayList.size() != 0) {
             for (int i = 0; i < MainActivity.bottomSheetPlayList.size(); i++) {
                 if (MainActivity.bottomSheetPlayList.get(i).getPnp().equals(storageManageDialogItem.getPnp())) {
@@ -118,7 +140,7 @@ public class StorageManageAdapter extends  RecyclerView.Adapter<StorageManageAda
         StorageManageDialog.progressBar.setVisibility(View.GONE);
     }
 
-    private void stopServiceWhenPlaylistZero(Context context) {
+    public static void stopServiceWhenPlaylistZero(Context context) {
         if (MainActivity.bottomSheetPlayList.size() == 0) {
             MainActivity.pands.setBackgroundResource(R.drawable.bottom_sheet_play);
             if (NotificationService.isPlaying) {
@@ -127,19 +149,28 @@ public class StorageManageAdapter extends  RecyclerView.Adapter<StorageManageAda
         }
     }
 
-    private void changePageItemBackground(int page, int position) {
-        if (page == 1) {
+    public static void changePageItemBackground(int page, int position) {
+        if (page == 1 && RainPage.arrayList.size() != 0) {
             RainPage.arrayList.get(position - 1).setIsplay(1);
             RainPage.adapter.notifyItemChanged(position - 1);
 //            Page1.adapter.notifyDataSetChanged();
-        } else if (page == 2) {
+        } else if (page == 2 && WaterPage.arrayList.size() != 0) {
+            WaterPage.arrayList.get(position - 1).setIsplay(1);
+            WaterPage.adapter.notifyItemChanged(position - 1);
+//            Page2.adapter.notifyDataSetChanged();
+        } else if (page == 3 && WindPage.arrayList.size() != 0) {
             WindPage.arrayList.get(position - 1).setIsplay(1);
             WindPage.adapter.notifyItemChanged(position - 1);
-//            Page2.adapter.notifyDataSetChanged();
-        } else if (page == 3) {
+        } else if (page == 4 && NaturePage.arrayList.size() != 0) {
+            NaturePage.arrayList.get(position - 1).setIsplay(1);
+            NaturePage.adapter.notifyItemChanged(position - 1);
+        } else if (page == 5 && ChakraPage.arrayList.size() != 0) {
             ChakraPage.arrayList.get(position - 1).setIsplay(1);
             ChakraPage.adapter.notifyItemChanged(position - 1);
-        } else if (page == 4) {
+        } else if (page == 6 && MantraPage.arrayList.size() != 0) {
+            MantraPage.arrayList.get(position - 1).setIsplay(1);
+            MantraPage.adapter.notifyItemChanged(position - 1);
+        } else if (page == 7 && HzPage.arrayList.size() != 0) {
             HzPage.arrayList.get(position - 1).setIsplay(1);
             HzPage.adapter.notifyItemChanged(position - 1);
         }
